+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+##### Changed
+* 💥 **\[arez\]** Upgrade Arez to version `0.39`.
+
 ### [v0.33](https://github.com/react4j/react4j/tree/v0.33) (2018-01-09)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.32...v0.33)