+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ##### Changed
 * 💥 **\[arez\]** Upgrade Arez to version `0.40`.
 * 💥 **\[annotations\]** Rename the `ReactComponent.Feature` constants to improve clarity;
-  - `TRUE` renamed to `ENABLED`
-  - `FALSE` renamed to `DISABLED`
+  - `TRUE` renamed to `ENABLE`
+  - `FALSE` renamed to `DISABLE`
   - `IF_DETECTED` renamed to `AUTODETECT`
 
 ### [v0.33](https://github.com/react4j/react4j/tree/v0.33) (2018-01-09)